Move HasVersionColumn to `entity` package
