+++ server/src/main/java/io/spine/server/projection/Projection.java
@@ -24,8 +24,8 @@ import com.google.protobuf.Message;
 import io.spine.annotation.Internal;
 import io.spine.core.Event;
 import io.spine.server.entity.EventPlayer;
+import io.spine.server.entity.HasVersionColumn;
 import io.spine.server.entity.TransactionalEntity;
-import io.spine.server.entity.storage.HasVersionColumn;
 import io.spine.server.event.EventSubscriber;
 import io.spine.server.projection.model.ProjectionClass;
 import io.spine.server.type.EventEnvelope;
